+++ lib/openapi_first/request_validation/failure.rb
@@ -3,10 +3,9 @@
 module OpenapiFirst
   module RequestValidation
     ## A failure object is returned when a request is invalid.
-    # It contains the status code, the location of the error and a message.
-    # It can be raised with #raise! to raise an error with a message.
     # @location [Symbol] can be one of :not_found, :body, :query, :header, :path, :cookie
-    # @status [Number] is the HTTP status code that fits the failure. This usually is 400, but can be 404, 405 or 415 as well.
+    # @status [Number] is the HTTP status code that fits the failure.
+    #         This usually is 400, but can be 404, 405 or 415 as well.
     # @message [String] is a generic error message
     # @validation_result [OpenapiFirst::Schema::ValidationResult] is the result of the JSON Schema validation
     class Failure
@@ -35,6 +34,7 @@ module OpenapiFirst
         @message || validation_result&.message || Rack::Utils::HTTP_STATUS_CODES[status]
       end
 
+      # Raise an exception that fits the failure.
       def raise!
         raise NotFoundError, error_message if location == :not_found